+++ docs/HISTORY.rst
@@ -11,6 +11,8 @@ Changelog
 - Remove unused ProjectsArea
   [cekk]
 
+- If a user is removed from a group is also removed from the notification group. ticket #10347
+  [arsenico13]
 
 2.1.3 (2015-09-22)
 ------------------